test: revert renaming of commonTests.ts and add bunfig.toml to ignore it from test coverage
