+++ bunfig.toml
@@ -1,0 +1,5 @@
+[test]
+coverage = true
+coveragePathIgnorePatterns = [
+  "tests/converters/helpers/commonTests.ts"
+]
