+++ tests/converters/vips.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, expect, test } from "bun:test";
 import type { ExecFileException } from "node:child_process";
 import { ExecFileFn } from "../../src/converters/types";
 import { convert } from "../../src/converters/vips";
-import { runCommonTests } from "./helpers/commonTests.test";
+import { runCommonTests } from "./helpers/commonTests";
 
 let calls: string[][] = [];
